Move md/barostat/utils to md/utils, update references
