+++ attic/barostat/test_barostat_water_density.py
@@ -14,7 +14,7 @@ from timemachine.fe.free_energy import AbsoluteFreeEnergy
 
 from md.ensembles import PotentialEnergyModel, NPTEnsemble
 from md.barostat.moves import MonteCarloBarostat
-from md.barostat.utils import get_bond_list, get_group_indices
+from md.utils import get_bond_list, get_group_indices
 from md.states import CoordsVelBox
 from md.utils import simulate_npt_traj
 from md.thermostat.moves import UnadjustedLangevinMove
