+++ md/ensembles.py
@@ -1,7 +1,7 @@
 import numpy as np
 from jax import numpy as jnp
 from simtk import unit
-from md.barostat.utils import compute_box_volume
+from md.utils import compute_box_volume
 from timemachine.constants import kB, ENERGY_UNIT, DISTANCE_UNIT
 
 from typing import Union
