+++ md/minimizer.py
@@ -7,7 +7,7 @@ from timemachine.lib import LangevinIntegrator, MonteCarloBarostat, custom_ops
 from ff.handlers import openmm_deserializer
 from ff import Forcefield
 from timemachine.fe.utils import get_romol_conf
-from md.barostat.utils import get_group_indices, get_bond_list
+from md.utils import get_group_indices, get_bond_list
 
 from rdkit import Chem
 from simtk import openmm
