+++ md/moves.py
@@ -5,7 +5,7 @@ from scipy.special import logsumexp
 
 from jax.scipy.special import logsumexp as jlogsumexp
 from timemachine.lib import custom_ops
-from md.barostat.utils import get_group_indices, get_bond_list
+from md.utils import get_group_indices, get_bond_list
 from timemachine import lib
 import jax
 import jax.numpy as jnp
