+++ slow_tests/test_benchmark.py
@@ -19,7 +19,7 @@ from timemachine.fe import free_energy
 from timemachine.fe.topology import SingleTopology
 
 from md import builders, minimizer
-from md.barostat.utils import get_bond_list, get_group_indices
+from md.utils import get_bond_list, get_group_indices
 
 
 def recenter(conf, box):
